Rename experiment-schema-types.ts to experiment-types.ts
