+++ src/mobile-app/components/app.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import { useState } from "react";
 import { ExperimentPicker } from "./experiment-picker";
 import { Experiment } from "../../shared/components/experiment";
-import { IExperiment } from "../../shared/experiment-schema-types";
+import { IExperiment } from "../../shared/experiment-types";
 
 import css from "./app.module.scss";
